latest changes and improvement in memory usage

+++ ERM_backend/src/main/java/ermorg/erm/dto/response/RiskResponse.java
@@ -2,13 +2,8 @@ package ermorg.erm.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import ermorg.erm.constant.BusinessVertical;
-import ermorg.erm.constant.Functions;
 import ermorg.erm.constant.RiskCategory;
-import ermorg.erm.constant.RiskSubCategory;
 import ermorg.erm.model.Risk;
-
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -22,8 +17,8 @@ public class RiskResponse {
     private RiskCategory category;
     private String subCategory;
     private String exposure;
-    private int function;
-    private int businessVertical;
+    private Long function;
+    private Long businessVertical;
     private String businessSegment;
     private long riskOwnerId;
     private long riskChampionId;
